+++ asabiris/exception_handler/email_exception_handler/__init__.py
@@ -1,7 +1,4 @@
 from .email_exception_handler import EmailExceptionHandlingStrategy
-from .api_exception_handler import APIExceptionHandlingStrategy
-
 __all__ = [
 	"EmailExceptionHandlingStrategy",
-	"APIExceptionHandlingStrategy"
 ]
