+++ asabiris/exception_handler/email_exception_handler/email_exception_handler.py
@@ -1,4 +1,4 @@
-from asabiris.exception_handler.exception_handler import ExceptionHandlingStrategy
+from asabiris.exception_handler_abc import ExceptionHandlingStrategy
 
 import logging
 
@@ -9,11 +9,11 @@ class EmailExceptionHandlingStrategy(ExceptionHandlingStrategy):
     def __init__(self, _, email_Service):
         self.EmailFailsafeManager = email_Service
 
-    async def handle_exception(self, exception, context_strategy):
+    async def handle_exception(self, exception, notification_params):
 
         # Extract 'from_email' and 'to_emails' from the context
-        from_email = context_strategy.get('from_email')
-        to_emails = context_strategy.get('to_emails')
+        from_email = notification_params.get('from_email')
+        to_emails = notification_params.get('to_emails')
 
         # Send error notification via Email Failsafe Manager
         if from_email and to_emails:
