--- asabiris/exception_handler/email_strategy/__init__.py
@@ -1,4 +1,0 @@
-from .email_exception_handling_strategy import EmailExceptionHandlingStrategy
-__all__ = [
-	"EmailExceptionHandlingStrategy",
-]
