+++ asabiris/exception_handler_abc.py
@@ -1,6 +1,7 @@
 from abc import ABC, abstractmethod
 
+
 class ExceptionHandlingStrategy(ABC):
     @abstractmethod
-    async def handle_exception(self, exception, context_strategy):
+    async def handle_exception(self, exception, notification_params):
         pass
