+++ asabiris/failsafe_abc.py
@@ -1,0 +1,7 @@
+from abc import ABC, abstractmethod
+
+
+class FailsafeHandler(ABC):
+    @abstractmethod
+    async def send_error_notification(self, error, sending_info):
+        pass
